remove duplicated tests, improve coverage
